+++ types/global.d.ts
@@ -7,3 +7,5 @@ declare module 'style-parser' {
 }
 
 declare const __DEV__: boolean;
+declare const __webpack_require__: Function;
+declare const __non_webpack_require__: Function;