+++ packages/emoji/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Added an `avoidFetch` option to `useEmojiData`.
 
+#### 🐞 Fixes
+
+- Fixed `emojibase-test-utils/test-data.json` being pulled into a Webpack bundle on accident.
+
 # 5.0.0 - 2019-10-29
 
 #### 💥 Breaking